servicedb/requirements.txt
pika
psycopg2-binary
psycopg2_pool
